+++ assets/init_shim.erl
@@ -1,0 +1,6 @@
+-module(init_shim).
+
+-export([start/0]).
+
+start() ->
+    init:boot().