+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.4] (unreleased)
+
+- Added support for the `--application` (or `-a`) option to support AtomVM OTP applications.
+
+### Changed
+
+- Using the `-s init` option is still supported but deprecated.  Use the `--application` (or `-a`) option to generate OTP applications using AtomVM.
+
 ## [0.7.3] (2023.11.25)
 
 - Added support for compiling "bootstrap" erlang files that `rebar3` otherwise cannot compile.